Add support for a default _template.js (#262)
Rather than writing a test targeting that a default template works,
I opted to remove all of the templates from the test fixtures. This
reduces the noise in the tests, and demonstrates that this feature
works.

There is curently a message to let the user know that they are using the
default template, and how to override it with their own custom
functionality.