Add stylus support (#1437)
* Add jest-cli to dependencies

I was getting a module resolution error after installing. Test command was
throwing an error saying it couldn't find jest-cli so now its in deps.

* Add stylus plugin

* Add support for using stylus modules with the plugin

* Add a validation whitelist to allow the "stylus" key in webpack config

This is a special case but we may run in to other cases before upgrading to
webpack v3. In webpack v1 many plugins and loaders used top-level configuration
keys to handle custom configuration. webpack-validator does not like this, but
it does provide the option to whitelist keys, which is exactly what I did here.

I also added a (hopefully) helpful error message to clarify what went wrong in
the case of an invalid webpack config.

* Add example of stylus plugin usage to gatsby-plugin-stylus

* Do not validate stylus options object shape

Validating the shape felt like overkill, and it might prevent valid options form
being used if I missed something.

* Configure stylus to use module extension for overall gatsby consistency

* Update readme to reflect css module configuration change

* Add common-tags helper to improve long string formatting

* Remove debug code