Update benchmark running instructions. (#9621)
<!--
  Q. Which branch should I use for my pull request?
  A. Use `master` branch (probably).

  Q. Which branch if my change is a bug fix for Gatsby v1?
  A. In this case, you should use the `v1` branch

  Q. Which branch if I'm still not sure?
  A. Use `master` branch. Ask in the PR if you're not sure and a Gatsby maintainer will be happy to help :)

  Note: We will only accept bug fixes for Gatsby v1. New features should be added to Gatsby v2.

  Learn more about contributing: https://www.gatsbyjs.org/docs/how-to-contribute/
-->

Add some details on how to run the benchmarks (create-pages, and markdown). I was looking for performance comparisons for a node optimization and it was not initially obvious how to run it. I since figured it out.  Here for example is a 3% performance improvement through the optimization in node.

NEW
```
success Building static HTML for pages — 0.613 s — 5001/5001 10403.89 pages/second
info Done building in 10.092 sec
```
OLD
```
success Building static HTML for pages — 0.631 s — 5001/5001 10071.25 pages/second
info Done building in 10.353 sec
```