fix(gatsby-plugin-sharp): use actions provided, don't assume Gatsby module location. (#9986)
The gatsby module resolve by require might not be the gatsby module that is executing. This can occur in mono-repo configurations.

This will prevent the Redux store from receiving the job actions and the bootstrap process will complete without waiting for all images to be processed. This can cause the build to finish without all images being generated.

Fixes #9984.