docs(categories): Add categories for Gallery and Infinite Scroll (#16594)
* Add categories

I recently created a starter and noticed that none of the current tags are suitable for it. But tags are required and it's not possible to add a starter without marking tags to it (even if it was possible, people wouldn't find it without tags). This is why I would like to add new tags "Infinite Scroll" and "Photos". I think both of these tags are fairly generic and applicable to many other projects besides my starter.

The starter is here: https://github.com/baobabKoodaa/gatsby-starter-photo-book

* Update categories.yml